Renames Kafka autoconfig types for easier conditions checks
