+++ zipkin-autoconfigure/collector-kafka/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-zipkin.autoconfigure.collector.kafka10.ZipkinKafka10CollectorAutoConfiguration
+zipkin.autoconfigure.collector.kafka10.ZipkinKafkaCollectorAutoConfiguration
