+++ zipkin-autoconfigure/collector-kafka08/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-zipkin.autoconfigure.collector.kafka.ZipkinKafkaCollectorAutoConfiguration
+zipkin.autoconfigure.collector.kafka.ZipkinKafka08CollectorAutoConfiguration
